'piyushnew'

+++ recommendation.zip_expanded/recommendation/src/main/java/com/ness/recommendation/model/PersonalLoan.java
@@ -27,5 +27,5 @@ public class PersonalLoan {
 	int minSalaryForSalaried;
 	int minSalaryForSelfEmployed;
 	int minCreditScore;
-	int contact;
+	double contact;
 }
